requirements.txt
Flask==0.12.1
requests >= 2.18.4
bs4 >= 0.0.1
python-firebase >= 1.2
PyDrive >= 1.3.1
PyPDF2 >= 1.26.0
